Adding ValueHolder superclass and updating relevant Parts
According to the original HyperTalk spec, certain Parts can have a
"value". Typically this means a mapping of the value of the Part to
one of its own properties. For example, the "value" of a Field is the
text content property of that field, while the value of a Button is
the title of the button as a string.

This commit introduces a new Part superclass that specifies a `.value`
getter and setter on all subclassed objects. This will allow us to
have a uniform method for ascertaining the value if a Part in
SimpleTalk statements like:
    put "hello" into field 2
Or:
    first world of field 1

The change will be useful when composing chunking statements