Alternate chunking pattern. String and Array implementations
With this branch/commit I am providing an alternate pattern for the
chunking API and demonstrating how it might work using Arrays and
Strings as the initial examples.

The key difference between this method and the preceding one is that a
'chunker' on an object is not itself a method, but an object with
three required methods:
1. getAll -- returns a list/array of all the chunks of the specified
kind (ie, all lines of text in a 'line' chunking of a string)
2. getAt -- Returns a specific chunk at a given index of similar
chunks on the source object (ie, line 3 of string)
3. setAt -- Returns a *new copy of the original object* where the
given chunk at the specified index is changed to some new value the
consumer passes in (`ie set character 4 of text to "p"` becomes
something like "hello".chunkers.character.setAt(3, "p"), giving us a
new string of "helpo")

I am already convinced that each object needs to know how to "chunk"
itself, and should specify which types of chunks it knows how to make
of itself. Each will also need to know how to re-set its own value(s)
based on setting chunks of a given kind at a given index. Because
strings are immutable here in JS, we have to return a copy of the
object in setAt(). Should that behavior be universal? What if the
object being chunked is a more complex object -- do we really want to
copy the whole thing when setting just a small part of it?