Initial pass at the Chunk object method
In this implementation, we keep the `chunkers` property of any objects
that we want to be chunkable, but instead of returning the actual
values of the 'chunks' of itself, we expect chunker functions to
return collections of instances of Chunk(). Chunk is merely a
functional wrapper that contains a value property (ie, the thing
itself like a string) and methods for further chunking itself based on
what its value allows as chunkers.

Because everything in the Chunking world is now described in terms of
Chunk or arrays of Chunk, we introduce the new 'combine' method on
named chunkers, which should "do the opposite" of the 'chunk' method
-- in other words, given a chunk name and an array of Chunk instances,
this function will re-combine the chunks into a single Chunk whose
value property appropriately combines the constituent chunks. For
example, if we have an array of 3 Chunk objects whose value properties
are the characters 'c' 'a' and 't' respectively, and we call 'combine'
on a String's 'character' chunker with these chunks as arguments, we
will get a new Chunk in response whose value is the string 'cat' (ie
with the characters assembled).