[BUGFIX] Fixing bug where drawing tools would not toggle
-- The Issue
When the halo is opened on a DrawingView, clicking the wrench button
was not toggling the drawing tools.

The expected behavior is as follows:
1. Clicking the wrench 'toggles' the model's "mode" part property to
either 'viewing' or 'drawing', depending on what it already has;
2. The prop change is handled by the view, which sets its element's
'mode' attribute to whatever the model's value has changed to (using
the new onPropChange method);
3. CSS is setup so that the element's 'mode' attribute values will
display or not display the painting tools. The presence of this
element attribute is also used in determining whether or not we can
draw on the image itself

The bug was a leftover call to `this.setAttribute()` present in the
overall webcomponent element's `onClick` handler. It was extremely
annoying to track down, but now everyting works

-- Additionally:
We have also updated the handler for the add drawing button in the
toolbox to automaticall put the new drawing into "drawing" mode when
the user uses this button to make a new drawing. Otherwise all
displayed/added drawings will not be in drawing mode by default when
first inserted into the DOM.