Switching mouseUp to click
-- About
We have been having some issues using the old Hypercard style
'mouseUp' as our basic button interaction handler. Drags, movement,
and other operations that could potentially end with a mouseUp over a
button will trigger click-like behavior.

Hypercard did not have a click handler initially, so mouseUp was the
only real option. Since we are working in the web world (and therefore
have click available out of the box) we should switch to using click.

This PR simply replaces all of the implemented mouseUp scripts and
handlers with click instead. We have also made minor changes to
ButtonView in order to map the correct events to the correct message
sends.

From now on, use `on click` instead of `on mouseUp`, unless you truly
want mouseUp.