Adding ExecutionContext class and updating code
-- About
This commit introduces a replacement for the basic _executionContext
object that had been added to Part instances in order to store
variable contexts. These contexts are used to dynamically lookup
variable bindings at runtime during the executiong of a message
handler on a given part.

However, this context object was not making a distinction between
local variable names in different handlers of the same script, meaning
one could overwrite a variable value in handler execution without
realizing it.

The new ExecutionContext object takes care of this by internally
storing variable bindings by message handler name. It provides an easy
API for looking up and setting local variable values (getLocal and
setLocal), and also provides the ability to change context, ie which
message handler is currently running the execution.

Added some integration tests that are passing.