[WIP] Adding targeting via Halo actions. Not yet working.
-- What
This is an initial commit of a new halo button and activity that
allows the user the interactively select the target for a given Part.

The targeting button on the Halo for a Part, when clicked, will
activate "targeting mode". Hovering over any other parts will make
them red (so you know what you might set as the target), and clicking
on one of these red possibilities will set the target for the original
part to the clicked part.

Right now there are some scripting errors when I try to make a good
example so we need to sort those out. But the general eventing and
property setting seems to work.