[WIP] Adjusting skin var/styler relationship
-- What
This commit introduces a new helper (colorToRGBList) as well as a new
expectation for the behavior of the `background-color` and
`background-transparency` properties.

First, each of these will overwrite a corresponding skin variable
value, `--skin-background-transparency` for the former and
`--skin-background-color-rgb` for the latter. Note that the latter
expects not the string beginning with `rgb()` or `rgba()`, but rather
a simple string of rgb values separated by commas (simply "255, 255,
255" for example). The CSS variables in the skin stylesheet are then
used to dynamically compute the actual rgba background color.

Setting either of these values to "default" will remove the
corresponding override of the variable from the cssStyle dict, and
therefore "fall back" to however the skin defines the variable for
that element.

The helper function `colorToRGBList` can take a string of `rgb(...)`
or `rgba(...)` and pick out the rgb values, formatting them into a
string that is acceptable as the `--skin-XXX-color-rgb` value.