[WIP] Adding more comprehensive skin object and lifecycle for it
-- What
[Note that this is not yet working]
We have added the 'skin' dyna-prop to Stack, which allows it to
recursively update any subparts. We have added a corresponding method
on the base Part class called `setStyleFromSkinDef` which takes a skin
object (as found by a Stack's dyna-prop lookup) and sets any specified
styles.

Right now this is not working because Stacks will try to set any props
they have to be the same on all Cards, but in this paradigm Cards do
not have skins, only stacks. This results in a property error.